+++ mocket/__init__.py
@@ -1,17 +1,17 @@
 from mocket.async_mocket import async_mocketize
+from mocket.compat import FakeSSLContext
 from mocket.core.socket import MocketSocket
+from mocket.core.ssl.context import MocketSSLContext
+from mocket.core.ssl.socket import MocketSSLSocket
 from mocket.entry import MocketEntry
 from mocket.mocket import Mocket
 from mocket.mocketizer import Mocketizer, mocketize
-from mocket.ssl.context import MocketSSLContext
-
-# NOTE this is here for backwards-compat to keep old import-paths working
-from mocket.ssl.context import MocketSSLContext as FakeSSLContext
 
 __all__ = [
     "Mocket",
     "MocketEntry",
     "MocketSSLContext",
+    "MocketSSLSocket",
     "MocketSocket",
     "Mocketizer",
     "async_mocketize",
