+++ mocket/compat/__init__.py
@@ -1,0 +1,5 @@
+from mocket.core.ssl.context import MocketSSLContext as FakeSSLContext
+
+__all__ = [
+    "FakeSSLContext",
+]
