+++ mocket/core/ssl/context.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Any
 
 from mocket.core.socket import MocketSocket
-from mocket.ssl.socket import MocketSSLSocket
+from mocket.core.ssl.socket import MocketSSLSocket
 
 
 class _MocketSSLContext:
